CP-16259: Make gfs2 sr layer backend agnostic
Created a new library - libiscsi - for all iscsi specific
functions. GFS2 SR code now determines backend kind based on URI
