+++ volume/org.xen.xapi.storage.gfs2/sr.py
@@ -9,13 +9,12 @@ import urlparse
 import xapi.storage.api.volume
 from xapi.storage.common import call
 from xapi.storage.libs import libvhd
+from xapi.storage.libs import libiscsi
 from xapi.storage import log
-import time
-import stat
-import XenAPI
 import fcntl
 import json
 import xcp.environ
+import XenAPI
 
 # For a block device /a/b/c, we will mount it at <mountpoint_root>/a/b/c
 mountpoint_root = "/var/run/sr-mount/"
@@ -24,7 +23,7 @@ def getSRpath(dbg, sr, check=True):
     uri = urlparse.urlparse(sr)
 
     if uri.scheme == 'iscsi':
-        (target, iqn, lunid) = decomposeISCSIuri(dbg, uri)
+        (target, iqn, lunid) = libiscsi.decomposeISCSIuri(dbg, uri)
         sr_path = "/dev/iscsi/%s/%s:3260/LUN%s" % (iqn, target, lunid)
     else:
         sr_path = uri.path
@@ -122,178 +121,27 @@ def umount(dbg, mnt_path):
     cmd = ["/usr/bin/umount", mnt_path]
     call(dbg, cmd)
 
-def login(dbg, target, iqn, usechap=False, username=None, password=None):
-    cmd = ["/usr/sbin/iscsiadm", "-m", "discovery", "-t", "st", "-p", target]
-    output = call(dbg, cmd).split('\n')[0] # FIXME: only take the first one returned. This might not always be the one we want.
-    log.debug("%s: output = %s" % (dbg, output))
-    portal = output.split(' ')[0]
-    # FIXME: error handling
-
-    # Provide authentication details if necessary
-    if usechap:
-        cmd = ["/usr/sbin/iscsiadm", "-m", "node", "-T", iqn, "--portal", portal, "--op", "update", "-n", "node.session.auth.authmethod", "-v", "CHAP"]
-        output = call(dbg, cmd)
-        log.debug("%s: output = %s" % (dbg, output))
-        cmd = ["/usr/sbin/iscsiadm", "-m", "node", "-T", iqn, "--portal", portal, "--op", "update", "-n", "node.session.auth.username", "-v", username]
-        output = call(dbg, cmd)
-        log.debug("%s: output = %s" % (dbg, output))
-        cmd = ["/usr/sbin/iscsiadm", "-m", "node", "-T", iqn, "--portal", portal, "--op", "update", "-n", "node.session.auth.password", "-v", password]
-        output = call(dbg, cmd)
-        log.debug("%s: output = %s" % (dbg, output))
-
-    # Log in
-    cmd = ["/usr/sbin/iscsiadm", "-m", "node", "-T", iqn, "--portal", portal, "-l"]
-    output = call(dbg, cmd)
-    log.debug("%s: output = %s" % (dbg, output))
-    # FIXME: check for success
-
-def waitForDevice(dbg):
-    # Wait for new device(s) to appear
-    cmd = ["/usr/sbin/udevadm", "settle"]
-    call(dbg, cmd)
-
-    # FIXME: For some reason, udevadm settle isn't sufficient to ensure the device is present. Why not?
-    time.sleep(10)
-
-def listSessions(dbg):
-    '''Return a list of (sessionid, portal, targetIQN) pairs representing logged-in iSCSI sessions.'''
-    cmd = ["/usr/sbin/iscsiadm", "-m", "session"]
-    output = call(dbg, cmd, error=False)  # if there are none, this command exits with rc 21
-    # e.g. "tcp: [1] 10.71.153.28:3260,1 iqn.2009-01.xenrt.test:iscsi6da966ca (non-flash)"
-    return [tuple([int(x.split(' ')[1].strip('[]')), x.split(' ')[2], x.split(' ')[3]]) for x in output.split('\n') if x <> '']
-
-def findMatchingSession(dbg, target, iqn, sessions):
-    for (sessionid, portal, targetiqn) in sessions:
-        # FIXME: only match on target IP address and IQN for now (not target port number)
-        if portal.startswith(target + ":") and targetiqn == iqn:
-            return sessionid
-    return None
-
-def rescanSession(dbg, sessionid):
-    cmd = ["/usr/sbin/iscsiadm", "-m", "session", "-r", str(sessionid), "--rescan"]
-    output = call(dbg, cmd)
-    log.debug("%s: output = '%s'" % (dbg, output))
-    # FIXME: check for success
-
-def getDesiredInitiatorName(dbg):
-    # FIXME: for now, get this from xapi. In future, xapi will write this to a file we can read from.
-    inventory = xcp.environ.readInventory()
-    session = XenAPI.xapi_local()
-    session.xenapi.login_with_password("root", "")
-    this_host = session.xenapi.host.get_by_uuid(inventory.get("INSTALLATION_UUID"))
-    return session.xenapi.host.get_other_config(this_host)['iscsi_iqn']
-
-def setInitiatorName(dbg, iqn):
-    with open('/etc/iscsi/initiatorname.iscsi', "w") as fd:
-        fd.write('InitiatorName=%s\n' % (iqn))
-
-def getCurrentInitiatorName(dbg):
-    with open('/etc/iscsi/initiatorname.iscsi', "r") as fd:
-        lines = fd.readlines()
-        for line in lines:
-            if not line.strip().startswith("#") and "InitiatorName" in line:
-                return line.split('=')[1].strip()
-
-def restartISCSIDaemon(dbg):
-    cmd = ["/usr/bin/systemctl", "restart", "iscsid"]
-    call(dbg, cmd)
-
-def isISCSIDaemonRunning(dbg):
-    cmd = ["/usr/bin/systemctl", "status", "iscsid"]
-    (stdout, stderr, rc) = call(dbg, cmd, error=False, simple=False)
-    return rc == 0
-
-def configureISCSIDaemon(dbg):
-    # Find out what the user wants the IQN to be
-    iqn = getDesiredInitiatorName(dbg)
-
-    # Make that the IQN, if possible
-    if not isISCSIDaemonRunning(dbg):
-        setInitiatorName(dbg, iqn)
-        restartISCSIDaemon(dbg)
-    else:
-        cur_iqn = getCurrentInitiatorName(dbg)
-        if iqn != cur_iqn:
-            if len(listSessions(dbg)) > 0:
-                raise xapi.storage.api.volume.Unimplemented("Daemon running with sessions from IQN '%s', desired IQN '%s'" % (cur_iqn, iqn))
-            else:
-                setInitiatorName(dbg, iqn)
-                restartISCSIDaemon(dbg)
-
-def zoneInLUN(dbg, uri):
-    log.debug("%s: zoneInLUN uri=%s" % (dbg, uri))
-
-    u = urlparse.urlparse(uri)
+def plug_device(dbg, uri):
+    u = urlparse.urlparse(uri)     
     if u.scheme == 'iscsi':
-        log.debug("%s: u = %s" % (dbg, u))
-        (target, iqn, lunid) = decomposeISCSIuri(dbg, u)
-        log.debug("%s: target = '%s', iqn = '%s', lunid = '%s'" % (dbg, target, iqn, lunid))
-
-        # If there's authentication required, the target will be of the form 'username%password@12.34.56.78'
-        atindex = target.find('@')
-        usechap = False
-        username = None
-        password = None
-        if atindex >= 0:
-            usechap = True
-            [username, password] = target[0:atindex].split('%')
-            target = target[atindex+1:]
-
-        configureISCSIDaemon(dbg)
-
-        current_sessions = listSessions(dbg)
-        log.debug("%s: current iSCSI sessions are %s" % (dbg, current_sessions))
-        sessionid = findMatchingSession(dbg, target, iqn, current_sessions)
-        if sessionid:
-            # If there's an existing session, rescan it in case new LUNs have appeared in it
-            log.debug("%s: rescanning session %d for %s on %s" % (dbg, sessionid, iqn, target))
-            # FIXME: should do refcounting to avoid logging out on first SR.detach
-            rescanSession(dbg, sessionid)
-        else:
-            # Otherwise, perform a fresh login
-            log.debug("%s: logging into %s on %s" % (dbg, iqn, target))
-            login(dbg, target, iqn, usechap, username, password)
-
-        waitForDevice(dbg)
-        dev_path = "/dev/iscsi/%s/%s:3260/LUN%s" % (iqn, target, lunid)
-    else:
+        dev_path = libiscsi.zoneInLUN(dbg, uri)
+    else: 
         # Assume it's a local block device
-        dev_path = u.path
-
-        if not(os.path.isdir(dev_path)) or not(os.path.ismount(dev_path)):
+        dev_path = "/%s%s" % (u.netloc, u.path)
+        # FIXME: Why do we need to check if it is a mount
+        # if not(os.path.exists(dev_path)) or not(os.path.ismount(dev_path)):
+        if not(os.path.exists(dev_path)):
             raise xapi.storage.api.volume.Sr_not_attached(dev_path)
-
-    # Verify it's a block device
-    if not stat.S_ISBLK(os.stat(dev_path).st_mode):
-        raise xapi.storage.api.volume.Unimplemented(
-            "Not a block device: %s" % dev_path)
-
-    # Switch to 'noop' scheduler
-    sched_file = "/sys/block/%s/queue/scheduler" % (os.path.basename(os.readlink(dev_path)))
-    with open(sched_file, "w") as fd:
-        fd.write("noop\n")
-
     return dev_path
 
-def zoneOutLUN(dbg, uri):
-    log.debug("%s: zoneOutLUN uri=%s" % (dbg, uri))
+def unplug_device(dbg, uri): 
+    u = urlparse.urlparse(uri)     
 
-    u = urlparse.urlparse(uri)
-    log.debug("%s: u = %s" % (dbg, u))
     if u.scheme == 'iscsi':
-        (target, iqn, lunid) = decomposeISCSIuri(dbg, u)
-        log.debug("%s: iqn = %s" % (dbg, iqn))
-
-        cmd = ["/usr/sbin/iscsiadm", "-m", "node", "-T", iqn, "-u"]
-        call(dbg, cmd)
-
-def decomposeISCSIuri(dbg, uri):
-    if (uri.scheme != "iscsi" or not uri.netloc or not uri.path):
-        raise xapi.storage.api.volume.SR_does_not_exist("The SR URI is invalid; please use iscsi://<target>/<targetIQN>/<lun>")
-
-    target = uri.netloc
-    [null, iqn, lunid] = uri.path.split("/")
-    return (target, iqn, lunid)
+        libiscsi.zoneOutLUN(dbg, uri)
+    else:
+        #do nothing for now
+        pass
 
 class Implementation(xapi.storage.api.volume.SR_skeleton):
 
@@ -318,7 +166,7 @@ class Implementation(xapi.storage.api.volume.SR_skeleton):
                     host, "gfs2setup", "gfs2Reload", {})
 
         # Zone in the LUN on this host
-        dev_path = zoneInLUN(dbg, uri)
+        dev_path = plug_device(dbg, uri)
 
         # Mount the gfs2 filesystem
         mnt_path = mount(dbg, dev_path)
@@ -344,7 +192,7 @@ class Implementation(xapi.storage.api.volume.SR_skeleton):
         fsname = "%s:%s" % (cluster_name, sr_name)
 
         # Zone-in the LUN
-        dev_path = zoneInLUN(dbg, uri)
+        dev_path = plug_device(dbg, uri)
         log.debug("%s: dev_path = %s" % (dbg, dev_path))
 
         # Make sure we wipe any previous GFS2 metadata
@@ -366,6 +214,7 @@ class Implementation(xapi.storage.api.volume.SR_skeleton):
         # Temporarily mount the filesystem so we can write the SR metadata
         mnt_path = mount_local(dbg, dev_path)
 
+        # FIXME: Move DB specific code to another place
         # create metadata DB
         import sqlite3
         conn = sqlite3.connect(mnt_path + "/sqlite3-metadata.db")
@@ -397,9 +246,7 @@ class Implementation(xapi.storage.api.volume.SR_skeleton):
             json_fp.write("\n")
 
         umount(dbg, mnt_path)
-
-        # Don't leave it zoned-in
-        zoneOutLUN(dbg, uri)
+        unplug_device(dbg, uri)
         return
 
     def destroy(self, dbg, sr):
@@ -412,11 +259,10 @@ class Implementation(xapi.storage.api.volume.SR_skeleton):
 
         # Unmount the FS
         sr_path = getSRpath(dbg, sr)
-        cmd = ["/usr/bin/umount", sr_path]
-        call(dbg, cmd)
+        umount(dbg, sr_path)
 
-        # Log out of the iSCSI session
-        zoneOutLUN(dbg, uri)
+        # Unplug device if need be
+        unplug_device(dbg, uri)
 
     def ls(self, dbg, sr):
         import volume
